refactor : backgroundSnow 한곳에 정의
- 눈 송이 생성도 같은 파일에서
